Always check if the GetLoop needs to be reset
When calling GetDone always check if the GetLoop needs to be reset. 
The reason I think this is reasonable is because if the server is hard closed we get the following error:
`GET: MQCC = MQCC_FAILED [2] MQRC = MQRC_CONNECTION_BROKEN [2009]`

When trying to gracefully handle the reconnection and then restart the `Get`, the poll is never restarted.